streamlit/requirements.txt
streamlit==1.37.1
joblib==1.4.2
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
requests==2.31.0
python-dotenv==1.0.1
